// src/store/types/login-types.ts
export interface LoginState {
    user: {
        created_at?: string
        email?: string
        email_verified_at?: string
        first_name?: string
        last_name?: string
        updated_at?: string
        uuid?: string
        __v?: number
        _id?: string
    },
    token: string
}